+++ src/components/OrderListItem.tsx
@@ -5,6 +5,7 @@ import { faBan, faCheck, faClock, faEllipsisH,
   faExclamationTriangle, faTimesCircle } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { hasChanged } from "../utils/hasChanged";
+import { orderExpirationTimestamp } from "../utils/misc";
 
 interface IOrderListItemProps {
   order: any;
@@ -25,14 +26,16 @@ export class OrderListItem extends React.Component<IOrderListItemProps, {}> {
       style += " order-list-item-selected";
     }
     let status = this.props.order.status;
-    const orderPlacedDate = moment(this.props.order.submitted_timestamp);
-    const orderExpirationDate = orderPlacedDate.clone().add(14, "days");
-    if (moment(orderExpirationDate).isBefore(moment.now())) {
+
+    // Make mock "expired" status based on completion date +2 weeks
+    const expirationTimestamp = orderExpirationTimestamp(this.props.order);
+    if (expirationTimestamp && expirationTimestamp.isBefore(moment.now())) {
       style += " order-list-item-expired";
       if (status === "complete" || status === "error" || status === "warning") {
         status = "expired";
       }
     }
+
     const delivery = (this.props.order.delivery === "ESI") ?
       "Zip" : this.props.order.delivery;
     const submitted = moment(this.props.order.submitted_timestamp).format(OrderListItem.timeFormat);