+++ CHANGELOG.md
@@ -1,6 +1,11 @@
-# v1.5.3 (unreleased)
+# v1.5.4 (unreleased)
 
 * Do not clip bounding box to spatial coverage unless bounding box is reset to global.
+
+# v1.5.3 (2020-06-17)
+
+* Show completed timestamp
+* Update expiration timestamp to completed + 2 weeks when available
 * Update dev hermes API URLs from `hermes2` to `hermes`
 * Use UAT CMR in staging only